cpp: Fix warnings, enforcing them
Fix the necessary warnings so that we can enable -Werror and stop
compilation at warnings.

+++ cpp/pr014.cpp
@@ -63,7 +63,7 @@ class Cached {
 };
 
 int Cached::chainLength(long long n) {
-  if (n < cache.size()) {
+  if (n < (long long)cache.size()) {
     auto sz = cache[n];
     if (sz >= 0)
       return sz;
